// src/screens/CoacheScreens/AuthScreens/CalendarScreen/user-list-screen/model/constructor-state.ts
import { UserArrays } from "../interface";

export interface StateInitial {
  type: "INITIAL";
  data: null;
  error: null;
}

export interface StateLoading {
  type: "LOADING";
  data: UserArrays | null;
  error: null;
}

export interface StateHasData {
  type: "HAS_DATA";
  data: UserArrays;
  error: null;
}

export interface StateEmpty {
  type: "EMPTY";
  data: Record<string, never>;
  error: null;
}

export interface StateError {
  type: "ERROR";
  data: null;
  error: Error;
}

export type UsersStates =
  | StateInitial
  | StateLoading
  | StateHasData
  | StateEmpty
  | StateError;